tibia-companion/requirements.txt
pymem>=1.13
websocket-client>=1.7.0
